backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ffmpeg-python>=0.2.0
psutil>=5.9.0
python-multipart>=0.0.6